feat: version 1.0

+++ qans/blog.py
@@ -20,7 +20,7 @@ def index():
     return render_template('blog/index.html', posts=posts)
 
 
-def get_post(id, check_author=True):
+def get_post(id):
     post = get_db().execute(
         'SELECT p.id, title, body, created, author_id, username'
         ' FROM post p JOIN user u ON p.author_id = u.id'
@@ -31,11 +31,20 @@ def get_post(id, check_author=True):
     if post is None:
         abort(404, "Post id {0} doesn't exist.".format(id))
 
-    if check_author and post['author_id'] != g.user['id']:
-        abort(403)
-
     return post
 
+def get_reply(id):
+    reply = get_db().execute(
+        'SELECT p.id, p.post_id, body, created, author_id, username'
+        ' FROM answers p JOIN user u ON p.author_id = u.id'
+        ' WHERE p.id = ?',
+        (id,)
+    ).fetchone()
+
+    if reply is None:
+        abort(404, "Reply id {0} doesn't exist.".format(id))
+
+    return reply
 
 @bp.route('/create', methods=('GET', 'POST'))
 @login_required
@@ -75,7 +84,7 @@ def postReply(id):
         (body, g.user['id'], id)
     )
     db.commit()
-    return redirect(url_for('blog.thread', id='id'))
+    return redirect(url_for('blog.thread', id=id))
 
 @bp.route('/<int:id>/update', methods=('GET', 'POST'))
 @login_required
@@ -100,7 +109,7 @@ def update(id):
                 (title, body, id)
             )
             db.commit()
-            return redirect(url_for('blog.index'))
+            return redirect(url_for('blog.thread', id=id))
 
     return render_template('blog/update.html', post=post)
 
@@ -108,21 +117,26 @@ def update(id):
 @bp.route('/<int:id>/updateReply', methods=('GET', 'POST'))
 @login_required
 def updateReply(id):
-    body = request.form['body']
-    postID = request.form['postID']
+    reply = get_reply(id)
+    error = None
+    if request.method == 'POST':
+        body = request.form['body']
+        postID = request.form['postID']
 
-    if not body:
-        error = 'Reply cannot be empty'
-    
-    if error is not None:
-        flash(error)
-    else:
-        db = get_db()
-        db.execute(
-            'UPDATE answers SET body=? WHERE id=?',
-            (body, id)
-        )
-        return redirect(url_for('thread', id=postID))
+        if not body:
+            error = 'Reply cannot be empty'
+        
+        if error is not None:
+            flash(error)
+        else:
+            db = get_db()
+            db.execute(
+                'UPDATE answers SET body=? WHERE id=?',
+                (body, id)
+            )
+            db.commit()
+            return redirect(url_for('blog.thread', id=postID))
+    return render_template('blog/updateReply.html', reply=reply)
 
 @bp.route('/<int:id>/thread', methods=('GET', ))
 @login_required
@@ -134,7 +148,7 @@ def thread(id):
         'SELECT p.id, body, created, author_id, post_id, username'
         ' FROM answers p JOIN user u ON p.author_id = u.id'
         ' WHERE p.post_id=?'
-        ' ORDER BY created DESC',
+        ' ORDER BY created ASC',
         (id, )
     ).fetchall()
     return render_template('blog/thread.html', post=post, replies=replies)
@@ -142,13 +156,22 @@ def thread(id):
 @bp.route('/<int:id>/delete', methods=('POST',))
 @login_required
 def delete(id):
-    """Delete a post.
-
-    Ensures that the post exists and that the logged in user is the
-    author of the post.
-    """
-    get_post(id)
+    post = get_post(id)
+    if g.user['id'] != post['author_id']:
+        abort(403)
     db = get_db()
     db.execute('DELETE FROM post WHERE id = ?', (id,))
+    db.execute('DELETE FROM answers WHERE post_id = ?', (id,))
     db.commit()
     return redirect(url_for('blog.index'))
+
+@bp.route('/<int:id>/deleteReply', methods=('POST',))
+@login_required
+def deleteReply(id):
+    reply = get_reply(id)
+    if g.user['id'] != reply['author_id']:
+        abort(403)
+    db = get_db()
+    db.execute('DELETE FROM answers WHERE id = ?', (id,))
+    db.commit()
+    return redirect(url_for('blog.thread', id=reply['post_id']))
